Initial commit.
Used the Hubble Console repository as a template. Removed the workspace pages, Redux integrations,
and other things. Right now, the console is not responsive to mobile screens. However, we intend
to make the website responsive to mobiles, which means we have to implement a drawer. For a desktop,
we should probably use a menu bar. For now, I think we can just use a drawer to save time.