Created the FAQs page.
The content used in this page is tentative. We need to review it before launching.

The template was created from Hubble Console. It appears that we forgot to remove the drawer width
from the main element, which added a margin of 60px on the left. It has been removed in this commit.